added authorization to the routes

+++ src/queries/ExamQueries.ts
@@ -1,3 +1,4 @@
+const asyncHandler = require('express-async-handler')
 import {
     GraphQLObjectType,
     GraphQLList,
@@ -20,16 +21,24 @@ const RootQuery = new GraphQLObjectType({
     fields: {
         exams: {
             type: new GraphQLList(ExamType),
-            async resolve(parent, args){
+            resolve: asyncHandler(async (parent:any, args:any, context:any) => {
+                const user = context.req.user;
+                if (!user) {
+                    throw new Error('You are not authorized')
+                }
                 return await Exam.find()
-            }
+            })
         },
         exam: {
             type: ExamType,
             args: { id:  { type: GraphQLID }},
-            async resolve(parent, args){
+            resolve: asyncHandler(async (parent:any, args:any, context:any) => {
+                const user = context.req.user;
+                if (!user) {
+                    throw new Error('You are not authorized')
+                }
                 return await Exam.findById(args.id)
-            }
+            })
         }
     }
 })
@@ -44,16 +53,26 @@ const Mutation = new GraphQLObjectType({
             args:{
                 name:{ type: new GraphQLNonNull(GraphQLString)  },
                 description:{ type: GraphQLString },
-                // userId: { type:  new GraphQLNonNull(GraphQLID)}
             },
-            async resolve(parent, args){
+            resolve: asyncHandler(async (parent:any, args:any, context:any) =>{
+                const user = context.req.user
+
+                if(!user){
+                    throw new Error('You are not authorized')
+                }
+
+                if(user.role !== 'admin'){
+                    throw new Error('You are not authorized. You are a student')
+                }
+
+                
                 const exam = await Exam.create({
                     name: args.name,
                     description: args.description,
-                    // userId: args.userId
+                    userId: user._id
                 }) as IExam
                 return exam
-            }
+            })
         },
         // Update An Exam
         updateExam: {
@@ -62,19 +81,34 @@ const Mutation = new GraphQLObjectType({
                 id: { type:  new GraphQLNonNull(GraphQLID)},
                 name:{ type: GraphQLString  },
                 description:{ type: GraphQLString },
-                // userId: { type:  new GraphQLNonNull(GraphQLID)}
             },
-            async resolve(parent, args) {
+            resolve: asyncHandler(async (parent:any, args:any, context:any) => {
+                const user = context.req.user
+
+                if(user.role !== 'admin'){
+                    throw new Error('Not Authorized')
+                }
+
+                const exam = Exam.findById({ id: args.id })
+
+                if (!exam) {
+                    throw new Error(`Exam with id ${args.id} does not exist`);
+                  }
+
+                if(exam.userId !== user._id){
+                    throw new Error('You are not authorized to update this exam')
+                }
+              
                 return await Exam.findByIdAndUpdate(
                     args.id, {
                         $set:{
                             name: args.name,
                             description: args.description,
-                            // userId: args.userId,
+                            userId: user._id,
                         }
                     }, { new: true }
                 )
-            }
+            })
         },
         // Delete An Exam
         deleteExam:{
@@ -82,9 +116,28 @@ const Mutation = new GraphQLObjectType({
             args:{
                 id: { type:  new GraphQLNonNull(GraphQLID)},
             },
-            async resolve(parent, args){
+            resolve: asyncHandler(async (parent:any, args:any, context:any) =>  {
+                const user = context.req.user
+
+                if(!user){
+                    throw new Error('You are not authorized')
+                }
+
+                if(user.role !== 'admin'){
+                    throw new Error('Not Authorized')
+                }
+
+                const exam = Exam.findById({ id: args.id })
+
+                if (!exam) {
+                    throw new Error(`Exam with id ${args.id} does not exist`);
+                  }
+
+                if(exam.userId !== user._id){
+                    throw new Error('You are not authorized to update this exam')
+                }
                 await Exam.findByIdAndRemove(args.id)
-            }
+            })
         }
     }
 })
