+++ src/queries/QAQueries.ts
@@ -1,3 +1,4 @@
+const asyncHandler = require('express-async-handler')
 import {
     GraphQLObjectType,
     GraphQLList,
@@ -9,9 +10,11 @@ import {
 } from 'graphql'
 
 import { ExamType } from '../schema/ExamSchema'
+const Exam = require('../models/ExamModel')
 import { QAType } from '../schema/QASchema'
 const QA = require('../models/QAModel')
 import { IQA } from '../models/QAModel'
+
 
 // Root Query
 const RootQuery = new GraphQLObjectType({
@@ -20,9 +23,11 @@ const RootQuery = new GraphQLObjectType({
         questions:{
             type: new GraphQLList(ExamType),
             args:{ examId: { type: GraphQLID }},
-            async resolve(parent, args) {
-                return await QA.find({ emailId: args.emailId })
-            }
+            resolve: asyncHandler(async (parent:any, args:any, context:any) =>  {
+                const user = context.req.user
+                if(!user) throw new Error('You are not authorized') 
+                return await QA.find({ examId: args.examId })
+            })
         }
     }
 })
@@ -35,6 +40,7 @@ const Mutation = new GraphQLObjectType({
         addQA:{
             type: QAType,
             args: {
+                examId: { type: new GraphQLNonNull(GraphQLID) },
                 question: { type: new GraphQLNonNull(GraphQLString)  },
                 optionOne: { type: GraphQLString },
                 optionTwo: { type: GraphQLString },
@@ -42,7 +48,16 @@ const Mutation = new GraphQLObjectType({
                 optionFour: { type: GraphQLString },
                 correctAnswer: { type: GraphQLInt },
             },
-            async resolve(parent, args) {
+            resolve: asyncHandler(async (parent:any, args:any, context:any) => {
+                const user = context.req.user
+                if(!user) throw new Error('You are not authorized')
+
+                if(user.admin === 'admin') throw new Error('You are a student. Error')
+
+                const exam = await Exam.findById({ id: args.examId})
+
+                if(!exam) throw new Error(`Exam with id ${args.id} does not exist`)
+
                 const qa = await QA.create({
                     question : args.question,
                     optionOne : args.optionOne,
@@ -50,9 +65,11 @@ const Mutation = new GraphQLObjectType({
                     optionThree : args.optionThree,
                     optionFour : args.optionFour,
                     correctAnswer : args.correctAnswer,
+                    examId: exam._id,
+                    userId: user._id
                 }) as IQA
                 return qa
-            }
+            })
         },
         // Update a QA
         updateQA:{
@@ -66,7 +83,18 @@ const Mutation = new GraphQLObjectType({
                 optionFour: { type: GraphQLString },
                 correctAnswer: { type: GraphQLInt },
             },
-            async resolve(parent, args){
+            resolve: asyncHandler(async (parent:any, args:any, context:any) => {
+                const user = context.req.user
+
+                if(!user) throw new Error('Not Authorized')
+                if(user.role !== 'admin') throw new Error('Not A Student')
+
+                const qa = await QA.findById({ id: args.id })
+                if(!qa) throw new Error(`Question with id of ${args.id} does not exist`)
+
+                if(qa.userId !== user._id)  throw new Error('You are not authorized to edit this question')
+
+
                 return await QA.findByIdAndUpdate(
                     args.id, {
                         $set:{
@@ -79,7 +107,7 @@ const Mutation = new GraphQLObjectType({
                         }
                     }, { new: true }
                 )
-            }
+            })
         },
         // Delete a QA
         deleteQA:{
@@ -87,9 +115,19 @@ const Mutation = new GraphQLObjectType({
             args: {
                 id: { type:  new GraphQLNonNull(GraphQLID)},
             },
-            async resolve(parent, args){
+            resolve: asyncHandler(async (parent:any, args:any, context:any) => {
+                const user = context.req.user
+
+                if(!user) throw new Error('Not Authorized')
+                if(user.role !== 'admin') throw new Error('Not A Student')
+
+                const qa = await QA.findById({ id: args.id })
+                if(!qa) throw new Error(`Question with id of ${args.id} does not exist`)
+
+                if(qa.userId !== user._id)  throw new Error('You are not authorized to edit this question')
+
                 await QA.findByIdAndRemove(args.id)
-            }
+            })
         }
     }
 })
